Day15 Solution

+++ src/Day03.kt
@@ -26,7 +26,7 @@ fun main() {
             var currentString = ""
             line.forEachIndexed { x, char ->
                 if (char.isDigit()) {
-                    currentNumber.add(Point2d(x, y))
+                    currentNumber.add(Point2d(x.toLong(), y.toLong()))
                     currentString += char
                 } else {
                     if (currentNumber.isNotEmpty()) {
